Update scripts.js
Updated the add() method , defined an array of expected keys containing the required keys for each pokemon.
Used the objectKeys method to take the array of the pokemon object and store it in a variable.
Used the if statement to check that the pokemon  parameter: is an object, ensure the length of the objectKeys array is equal to that of the expectedKeys, and then used the every() method to ensure that all of the objectKeys are included within the expectedKeys.